requirements.txt
fastapi
motor
python-multipart
python-jose
passlib
pydantic
email-validator
cryptography
python-dotenv
uvicorn
